// Capstone Project/src/app/pages/admin/admin-amenity-management-page/screen/types.ts
export enum adminPageAmenityEnum {
  GET_DATA_AMENITY_REQUEST = 'APP/ADMIN_PAGE/GET_DATA_AMENITY_REQUEST',
  GET_DATA_AMENITY_SUCCESS = 'APP/ADMIN_PAGE/GET_DATA_AMENITY_SUCCESS',
  GET_DATA_AMENITY_FAILURE = 'APP/ADMIN_PAGE/GET_DATA_AMENITY_FAILURE',

  CREATE_DATA_AMENITY_REQUEST = 'APP/ADMIN_PAGE/CREATE_DATA_AMENITY_REQUEST',
  CREATE_DATA_AMENITY_SUCCESS = 'APP/ADMIN_PAGE/CREATE_DATA_AMENITY_SUCCESS',
  CREATE_DATA_AMENITY_FAILURE = 'APP/ADMIN_PAGE/CREATE_DATA_AMENITY_FAILURE',

  UPDATE_DATA_AMENITY_REQUEST = 'APP/ADMIN_PAGE/UPDATE_DATA_AMENITY_REQUEST',
  UPDATE_DATA_AMENITY_SUCCESS = 'APP/ADMIN_PAGE/UPDATE_DATA_AMENITY_SUCCESS',
  UPDATE_DATA_AMENITY_FAILURE = 'APP/ADMIN_PAGE/UPDATE_DATA_AMENITY_FAILURE',

  DELETE_DATA_AMENITY_REQUEST = 'APP/ADMIN_PAGE/DELETE_DATA_AMENITY_REQUEST',
  DELETE_DATA_AMENITY_SUCCESS = 'APP/ADMIN_PAGE/DELETE_DATA_AMENITY_SUCCESS',
  DELETE_DATA_AMENITY_FAILURE = 'APP/ADMIN_PAGE/DELETE_DATA_AMENITY_FAILURE',

  CLEAR_AMENITY_STATE = 'APP/ADMIN_PAGE/CLEAR_AMENITY_STATE',
}